claudeflow/requirements.txt
# ClaudeFlow Dependencies

# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Async HTTP client
aiohttp==3.9.1
aiofiles==23.2.0

# Google APIs
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Configuration and environment
python-dotenv==1.0.0
pyyaml==6.0.1

# Retry logic and error handling
tenacity==8.2.3

# Logging and monitoring
structlog==23.2.0

# Database (Optional)
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0  # PostgreSQL async driver
aiosqlite==0.19.0  # SQLite async driver

# Caching and rate limiting (Optional)
redis==5.0.1
slowapi==0.1.9

# Security
cryptography==41.0.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Additional utilities
jinja2==3.1.2  # Template engine
python-multipart==0.0.6  # File uploads
httpx==0.25.2  # Alternative HTTP client

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Monitoring and error tracking (Optional)
sentry-sdk[fastapi]==1.38.0

# Email notifications (Optional)
aiosmtplib==3.0.1
email-validator==2.1.0

# File processing (Optional)
python-magic==0.4.27
pillow==10.1.0

# Webhook validation
hmac

# JSON processing
orjson==3.9.10

# Date/time utilities
python-dateutil==2.8.2

# Environment variable validation
pydantic-settings==2.0.3